Address some warnings

+++ cr_stack.h
@@ -166,7 +166,7 @@ int crcb_device_get_info(cr_DeviceInfoResponse *pDi);
 //*************************************************************************
 //  Link (ping) Service
 //*************************************************************************
-int crcb_ping_get_signal_strength(int32_t *rssi);
+int crcb_ping_get_signal_strength(int8_t *rssi);
 
 //*************************************************************************
 //  CLI Service
